Move `DisplayUnifiedInbox` and `FolderList` to ´:feature:navigation:drawer`
