+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/legacy/DisplayUnifiedInbox.kt
@@ -1,4 +1,4 @@
-package app.k9mail.legacy.ui.folder
+package app.k9mail.feature.navigation.drawer.legacy
 
 data class DisplayUnifiedInbox(
     val unreadMessageCount: Int,
